move wds loading to hook
