+++ src/workspace-data/Data.js
@@ -2,11 +2,10 @@ import { Interactive, Spinner } from '@terra-ui-packages/components';
 import FileSaver from 'file-saver';
 import _ from 'lodash/fp';
 import * as qs from 'qs';
-import { Fragment, useCallback, useEffect, useImperativeHandle, useMemo, useRef, useState } from 'react';
+import { Fragment, useCallback, useEffect, useImperativeHandle, useRef, useState } from 'react';
 import { DraggableCore } from 'react-draggable';
 import { div, form, h, h3, input, span } from 'react-hyperscript-helpers';
 import { cloudProviders } from 'src/analysis/utils/runtime-utils';
-import { authOpts } from 'src/auth/auth-session';
 import * as breadcrumbs from 'src/components/breadcrumbs';
 import Collapse from 'src/components/Collapse';
 import { ButtonOutline, Clickable, DeleteConfirmationModal, Link, spinnerOverlay } from 'src/components/common';
@@ -16,9 +15,8 @@ import { ConfirmedSearchInput } from 'src/components/input';
 import { MenuButton } from 'src/components/MenuButton';
 import { MenuDivider, MenuTrigger } from 'src/components/PopupTrigger';
 import { Ajax } from 'src/libs/ajax';
-import { fetchWDS } from 'src/libs/ajax/ajax-common';
 import { EntityServiceDataTableProvider } from 'src/libs/ajax/data-table-providers/EntityServiceDataTableProvider';
-import { resolveWdsApp, WdsDataTableProvider, wdsProviderName } from 'src/libs/ajax/data-table-providers/WdsDataTableProvider';
+import { wdsProviderName } from 'src/libs/ajax/data-table-providers/WdsDataTableProvider';
 import { appStatuses } from 'src/libs/ajax/leonardo/models/app-models';
 import colors from 'src/libs/colors';
 import { getConfig } from 'src/libs/config';
@@ -52,6 +50,7 @@ import { getReferenceData, getReferenceLabel } from './reference-data/reference-
 import { ReferenceDataContent } from './reference-data/ReferenceDataContent';
 import { ReferenceDataDeleter } from './reference-data/ReferenceDataDeleter';
 import { ReferenceDataImporter } from './reference-data/ReferenceDataImporter';
+import { useDataTableProvider } from './useDataTableProvider';
 import { WorkspaceAttributes } from './WorkspaceAttributes';
 
 const styles = {
@@ -343,7 +342,8 @@ const DataTableActions = ({
             },
             [input({ type: 'hidden', name: 'FCtoken', value: getTerraUser().token }), input({ type: 'hidden', name: 'model', value: 'flexible' })]
           ),
-          (dataProvider.features.supportsTsvDownload || dataProvider.features.supportsTsvAjaxDownload) &&
+          dataProvider &&
+            (dataProvider.features.supportsTsvDownload || dataProvider.features.supportsTsvAjaxDownload) &&
             h(
               MenuButton,
               {
@@ -371,7 +371,8 @@ const DataTableActions = ({
               },
               'Download TSV'
             ),
-          dataProvider.features.supportsExport &&
+          dataProvider &&
+            dataProvider.features.supportsExport &&
             h(
               MenuButton,
               {
@@ -388,7 +389,8 @@ const DataTableActions = ({
               },
               'Export to workspace'
             ),
-          dataProvider.features.supportsTypeRenaming &&
+          dataProvider &&
+            dataProvider.features.supportsTypeRenaming &&
             h(
               MenuButton,
               {
@@ -399,7 +401,8 @@ const DataTableActions = ({
               },
               'Rename table'
             ),
-          dataProvider.features.supportsTypeDeletion &&
+          dataProvider &&
+            dataProvider.features.supportsTypeDeletion &&
             h(
               MenuButton,
               {
@@ -571,11 +574,6 @@ export const WorkspaceData = _.flow(
     const [showDataTableVersionHistory, setShowDataTableVersionHistory] = useState({}); // { [entityType: string]: boolean }
     const pollWdsInterval = useRef();
 
-    const [wdsApp, setWdsApp] = useState({ status: 'None', state: undefined });
-    const [wdsTypes, setWdsTypes] = useState({ status: 'None', state: [] });
-    const [wdsCapabilities, setWdsCapabilities] = useState({ status: 'None', state: undefined });
-    const [useCwds, setUseCwds] = useState(undefined);
-
     const { dataTableVersions, loadDataTableVersions, saveDataTableVersion, deleteDataTableVersion, importDataTableVersion } =
       useDataTableVersions(workspace);
 
@@ -584,18 +582,10 @@ export const WorkspaceData = _.flow(
 
     const { runningJobs: runningImportJobs, refresh: refreshRunningImportJobs } = useImportJobs(workspace);
     const signal = useCancellation();
-    const cwdsURL = getConfig().cwdsUrlRoot;
 
     const entityServiceDataTableProvider = new EntityServiceDataTableProvider(namespace, name);
     const region = isAzureWorkspace ? storageDetails.azureContainerRegion : storageDetails.googleBucketLocation;
-
-    const wdsDataTableProvider =
-      !useCwds &&
-      useMemo(() => {
-        const app = wdsApp?.state;
-        const proxyUrl = app?.proxyUrls?.wds;
-        return new WdsDataTableProvider(workspaceId, proxyUrl, wdsCapabilities?.state);
-      }, [workspaceId, wdsApp, wdsCapabilities]);
+    const [wdsDataTableProvider, wdsApp, wdsTypes, setWdsTypes, loadWdsData] = useDataTableProvider(workspaceId);
 
     const loadEntityMetadata = async () => {
       try {
@@ -659,127 +649,6 @@ export const WorkspaceData = _.flow(
       }
     };
 
-    // returns the found app only if it is in a ready state,
-    // otherwise returns undefined
-    const loadWdsApp = useCallback((workspaceId) => {
-      return Ajax()
-        .Apps.listAppsV2(workspaceId)
-        .then((apps) => {
-          const foundApp = resolveWdsApp(apps);
-          switch (foundApp?.status) {
-            case appStatuses.provisioning.status:
-            case appStatuses.updating.status:
-              setWdsApp({ status: 'Loading', state: foundApp });
-              break;
-            case appStatuses.running.status:
-              setWdsApp({ status: 'Ready', state: foundApp });
-              return foundApp;
-            case appStatuses.error.status:
-              setWdsApp({ status: 'Error', state: foundApp });
-              break;
-            default:
-              if (foundApp?.status) {
-                // eslint-disable-next-line no-console
-                console.log(`Unhandled state [${foundApp?.status} while polling WDS`);
-              }
-          }
-        })
-        .catch((error) => {
-          setWdsApp({ status: 'Error', state: 'Error resolving WDS app' });
-          reportError('Error resolving WDS app', error);
-        });
-    }, []);
-
-    const loadWdsTypes = useCallback(
-      (url, workspaceId) => {
-        setWdsTypes({ status: 'None', state: [] });
-        return Ajax(signal)
-          .WorkspaceData.getSchema(url, workspaceId)
-          .then((typesResult) => {
-            setWdsTypes({ status: 'Ready', state: typesResult });
-          })
-          .catch((error) => {
-            setWdsTypes({ status: 'Error', state: 'Error loading WDS schema' });
-            reportError('Error loading WDS schema', error);
-          });
-      },
-      [signal]
-    );
-
-    const loadWdsCapabilities = useCallback(
-      async (url) => {
-        try {
-          const capabilitiesResult = await Ajax(signal).WorkspaceData.getCapabilities(url);
-          setWdsCapabilities({ status: 'Ready', state: capabilitiesResult });
-        } catch (error) {
-          setWdsCapabilities({ status: 'Error', state: 'Error loading WDS capabilities' });
-          reportError('Error loading WDS capabilities', error);
-        }
-      },
-      [signal]
-    );
-
-    const loadWdsData = useCallback(async () => {
-      // Try to load the proxy URL
-      if (!wdsApp || !['Ready', 'Error'].includes(wdsApp.status)) {
-        const foundApp = await loadWdsApp(workspaceId);
-        // TODO: figure out how not to make this redundant fetch, per
-        // https://github.com/DataBiosphere/terra-ui/pull/4202#discussion_r1319145491
-        if (foundApp) {
-          loadWdsTypes(foundApp.proxyUrls?.wds, workspaceId);
-          loadWdsCapabilities(foundApp.proxyUrls?.wds);
-        }
-      }
-      // If we have the proxy URL try to load the WDS types
-      else if (wdsApp?.status === 'Ready') {
-        const proxyUrl = wdsApp.state.proxyUrls?.wds;
-        await loadWdsTypes(proxyUrl, workspaceId);
-        await loadWdsCapabilities(proxyUrl);
-      }
-    }, [wdsApp, loadWdsApp, loadWdsTypes, loadWdsCapabilities, workspaceId]);
-
-    /// ////////////////// To be moved to a new file ///////////////////////
-
-    const loadCwdsTypes = useCallback(
-      (workspaceId) => {
-        setWdsTypes({ status: 'None', state: [] });
-        return Ajax(signal)
-          .WorkspaceData.getSchema(cwdsURL, workspaceId)
-          .then((typesResult) => {
-            setWdsTypes({ status: 'Ready', state: typesResult });
-          })
-          .catch((error) => {
-            setWdsTypes({ status: 'Error', state: 'Error loading WDS schema' });
-            reportError('Error loading WDS schema', error);
-          });
-      },
-      [signal, cwdsURL]
-    );
-
-    const loadCwdsCapabilities = useCallback(async () => {
-      try {
-        const capabilitiesResult = await Ajax(signal).WorkspaceData.getCapabilities(cwdsURL);
-        setWdsCapabilities({ status: 'Ready', state: capabilitiesResult });
-      } catch (error) {
-        setWdsCapabilities({ status: 'Error', state: 'Error loading WDS capabilities' });
-        reportError('Error loading WDS capabilities', error);
-      }
-    }, [signal, cwdsURL]);
-
-    const loadCwdsDataTableProvider =
-      useCwds &&
-      useMemo(() => {
-        return new WdsDataTableProvider(workspaceId, cwdsURL, wdsCapabilities?.state);
-      }, [workspaceId, cwdsURL, wdsCapabilities]);
-
-    const loadCwdsData = useCallback(async () => {
-      await loadCwdsTypes(workspaceId);
-      await loadCwdsCapabilities();
-      await loadCwdsDataTableProvider();
-    }, [loadCwdsDataTableProvider, loadCwdsTypes, loadCwdsCapabilities, workspaceId]);
-
-    /// ///////////////////////////////////////////////////////////////////
-
     const toSortedPairs = _.flow(_.toPairs, _.sortBy(_.first));
 
     const searchAcrossTables = async (typeNames, activeCrossTableTextFilter) => {
@@ -841,30 +710,7 @@ export const WorkspaceData = _.flow(
         setSnapshotMetadataError(false);
         setEntityMetadata({});
 
-        const checkCWDS = async () => {
-          try {
-            const response = await fetchWDS(cwdsURL)(`collections/v1/${workspaceId}`, _.merge(authOpts(), { signal }));
-            const data = await response.json();
-            if (Object.keys(data).length !== 0) {
-              // setWdsUrl({ status: 'Ready', state: cwdsURL });
-              setUseCwds(true);
-              setWdsApp({ status: 'Ready', state: undefined }); // No app needed for CWDS
-            } else {
-              // setWdsUrl({ status: 'Loading', state: null });
-              setUseCwds(false);
-            }
-          } catch (error) {
-            console.error(error);
-            // setWdsUrl({ status: 'Loading', state: null });
-            setUseCwds(false);
-          }
-        };
-
-        if (useCwds === undefined) {
-          checkCWDS();
-        } else if (useCwds) {
-          // Don't need to worry about WDS app
-        } else if (!wdsReady && !wdsError && !pollWdsInterval.current) {
+        if (!wdsReady && !wdsError && !pollWdsInterval.current) {
           // Start polling if we're missing WDS Types, and stop polling when we have them.
           pollWdsInterval.current = setInterval(loadWdsData, 30 * 1000);
         } else if (wdsReady && pollWdsInterval.current) {
@@ -876,7 +722,7 @@ export const WorkspaceData = _.flow(
           pollWdsInterval.current = undefined;
         };
       }
-    }, [loadWdsData, workspaceId, wdsApp, wdsTypes, isAzureWorkspace, signal, useCwds, wdsError, wdsReady, cwdsURL]);
+    }, [loadWdsData, workspaceId, wdsApp, wdsTypes, isAzureWorkspace, signal, wdsError, wdsReady]);
 
     const canUploadTsv = isGoogleWorkspace || (isAzureWorkspace && wdsReady);
     return div({ style: styles.tableContainer }, [
@@ -1533,7 +1379,6 @@ export const WorkspaceData = _.flow(
                 [
                   workspaceDataTypes.wds,
                   () =>
-                    wdsDataTableProvider &&
                     wdsReady &&
                     !_.isEmpty(wdsTypes.state) &&
                     h(WDSContent, {
